Split review command module into cohesive submodules
Slice 4 of #297. Break extensions/the-last-harness/review.ts
(1,369 lines) into focused sibling modules to improve reviewability,
while preserving the module's public API and behavior:

- review-args.ts: review-mode constants/types, dispatch/parsed-args
  types, decideBranchAction, tokenizeArgs, parseReviewArgs
- review-github.ts: GitHub PR ref resolution (URL/slug/remote
  parsing) and REST metadata/diff fetch helpers
- review-envelope.ts: ReviewGatheredContext, untracked delimiters,
  buildReviewEnvelope and the content-escaping/snapshot helpers
- review.ts (1,369 -> 883 lines): retains the interactive picker,
  gather implementations, dispatch, and command registration, and
  re-exports the previously public symbols so importers and tests are
  unaffected

Sibling modules are imported with explicit .ts extensions because
review.ts is loaded in tests via Node native strip-types (see the
issue #296 note in review.ts), which does not remap ".js" to ".ts";
tsconfig gains allowImportingTsExtensions to accept this under tsc.

Pure refactor: moved declarations are unchanged and the dependency
direction is one-way (no module imports review.ts). npm run validate
passes.
